% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.R
\name{dic}
\alias{dic}
\title{Deviance Information Criterion}
\usage{
dic(mod = NULL, form = NULL)
}
\description{
\code{dic} returns the deviance information criterion for the model
  and allows you to specify which DIC to return (based on the 8
  DIC equations for mixture models proposed by Celeux et al. 2006).
}
