% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Log-likelihood}
\usage{
loglik(Z = NULL, Y = NULL, X = NULL, disease = NULL,
  region = NULL, Beta = NULL, at_risk = NULL, phi = NULL,
  pi = NULL, alloc = NULL)
}
\description{
\code{loglik} calculates the log-likelihood component of the SPMM
}
