% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis-B.R
\name{metropolis_B}
\alias{metropolis_B}
\title{Metropolis B}
\usage{
metropolis_B(Z = NULL, Y = NULL, X = NULL, D = NULL, W = NULL,
  at_risk = NULL, disease = NULL, region = NULL, Sigma = NULL,
  v = NULL, k = NULL, B = NULL, B_scale = NULL, u_matrix = NULL,
  Beta = NULL, phi = NULL, pi = NULL)
}
\description{
Do thing for stuff
}
