% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis-beta.R
\name{metropolis_beta}
\alias{metropolis_beta}
\title{Metropolis Beta}
\usage{
metropolis_beta(Z = NULL, Y = NULL, X = NULL, at_risk = NULL,
  disease = NULL, region = NULL, Beta = NULL, beta_scale = NULL,
  phi = NULL, pi = NULL)
}
\description{
\code{metropolis_beta} runs the Metropolis-Hastings algorithm to
  update each Beta term one at a time
}
