% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis-pi.R
\name{metropolis_pi}
\alias{metropolis_pi}
\title{Metropolis pi}
\usage{
metropolis_pi(Y = NULL, Z = NULL, X = NULL, at_risk = NULL,
  disease = NULL, region = NULL, Beta = NULL, phi = NULL,
  pi = NULL, N = NULL)
}
\description{
Metropolis pi
}
