% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis-sigma.R
\name{metropolis_sigma}
\alias{metropolis_sigma}
\title{Metropolis sigma}
\usage{
metropolis_sigma(Z = NULL, Y = NULL, X = NULL, D = NULL,
  W = NULL, at_risk = NULL, disease = NULL, region = NULL,
  Sigma = NULL, sigma_scale = NULL, v = NULL, k = NULL, B = NULL,
  Beta = NULL, phi = NULL, u_matrix = NULL, pi = NULL)
}
\description{
Does the thing for sigma
}
