% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis-u.R
\name{metropolis_u}
\alias{metropolis_u}
\title{Metropolis u}
\usage{
metropolis_u(Z = NULL, Y = NULL, X = NULL, D = NULL, W = NULL,
  at_risk = NULL, disease = NULL, region = NULL, Sigma = NULL,
  v = NULL, k = NULL, B = NULL, B_scale = NULL, u_matrix = NULL,
  Beta = NULL, phi = NULL, pi = NULL, constant_B = FALSE)
}
\description{
Metropolis u
}
