% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_z.R
\name{metropolis_z}
\alias{metropolis_z}
\title{Metropolis Z}
\usage{
metropolis_z(Y = NULL, Z = NULL, alloc = NULL, X = NULL,
  at_risk = NULL, disease = NULL, region = NULL, Beta = NULL,
  phi = NULL, pi = NULL, N = NULL)
}
\description{
Does M-H algo for Z, the allocation vector
}
