% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmm.R
\name{spmm}
\alias{spmm}
\title{SPMM}
\usage{
spmm(data = NULL, inits = NULL, n_iter = NULL, n_burnin = NULL,
  n_chains = 3, constant_B = FALSE)
}
\arguments{
\item{data}{The actual data}

\item{n_iter}{How many iterations to sample?}

\item{n_burnin}{How many iterations to burn in chain?}

\item{n_chains}{How many chains to run per parameter?}

\item{C}{The number of diseases to model}

\item{R}{The number of spatial regions}

\item{D}{R x R neighbourhood matrix with entries D_jj, j = 1...R, i.e. the count of neighbours for region j}

\item{W}{R x R adjacency matrix}

\item{m}{Number of covariates}
}
\description{
\code{spmm} is the main function to run a spatial Poisson mixture model
}
